refactor[api-server]: lottery.service 패키지 위치 변경
