+++ src/main/java/com/watermelon/server/admin/controller/AdminExpectationController.java
@@ -4,7 +4,7 @@ package com.watermelon.server.admin.controller;
 import com.watermelon.server.admin.dto.response.ResponseAdminExpectationApprovedDto;
 import com.watermelon.server.event.lottery.dto.response.ResponseExpectationDto;
 import com.watermelon.server.event.lottery.error.ExpectationNotExist;
-import com.watermelon.server.lottery.service.ExpectationService;
+import com.watermelon.server.event.lottery.service.ExpectationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
