+++ src/test/java/com/watermelon/server/lottery/auth/interceptor/LoginCheckInterceptorTest.java
@@ -4,7 +4,7 @@ import com.watermelon.server.event.lottery.auth.exception.AuthenticationExceptio
 import com.watermelon.server.event.lottery.auth.interceptor.LoginCheckInterceptor;
 import com.watermelon.server.event.lottery.auth.service.TokenVerifier;
 import com.watermelon.server.event.lottery.link.service.LinkService;
-import com.watermelon.server.lottery.service.LotteryService;
+import com.watermelon.server.event.lottery.service.LotteryService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.assertj.core.api.Assertions;
