+++ src/test/java/com/watermelon/server/lottery/link/service/LinkServiceImplTest.java
@@ -5,7 +5,7 @@ import com.watermelon.server.event.lottery.domain.LotteryApplier;
 import com.watermelon.server.event.lottery.link.dto.MyLinkDto;
 import com.watermelon.server.event.lottery.link.repository.LinkRepository;
 import com.watermelon.server.event.lottery.link.service.LinkServiceImpl;
-import com.watermelon.server.lottery.service.LotteryService;
+import com.watermelon.server.event.lottery.service.LotteryService;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
